refactor(redux): remove actions and reducers
moved out some bits to:
- lib/types.ts
- utils/generateIdentitySummary.ts
