+++ package.json
@@ -79,8 +79,6 @@
     "i18n-js": "^3.1.0",
     "jolocom-lib": "^4.1.0",
     "ramda": "^0.26.1",
-    "redux": "^3.7.2",
-    "redux-thunk": "^2.3.0",
     "sjcl": "^1.0.7",
     "sqlite3": "^4.0.0",
     "squel": "^5.12.1",
