--- src/actions/account/actionTypes.ts
@@ -1,1 +1,0 @@
-export const HAS_EXTERNAL_CREDENTIALS = 'HAS_EXTERNAL_CREDENTIALS'
