--- src/actions/index.ts
@@ -1,17 +1,0 @@
-import * as accountActions from './account/'
-import * as registrationActions from './registration/'
-import * as navigationActions from './navigation/'
-import * as genericActions from './generic/'
-import * as ssoActions from './sso/'
-import * as recoveryActions from './recovery'
-import * as notificationsActions from './notifications'
-
-export {
-  accountActions,
-  registrationActions,
-  navigationActions,
-  genericActions,
-  ssoActions,
-  recoveryActions,
-  notificationsActions,
-}
