--- src/actions/registration/loadingStages.ts
@@ -1,9 +1,0 @@
-import I18n from '../../locales/i18n'
-import strings from '../../locales/strings'
-
-export const loadingStages = [
-  I18n.t(strings.ENCRYPTING_AND_STORING_DATA_LOCALLY),
-  I18n.t(strings.FUELING_WITH_ETHER),
-  I18n.t(strings.REGISTERING_DECENTRALIZED_IDENTITY),
-  I18n.t(strings.PREPARING_LAUNCH),
-]
