--- src/actions/sso/index.ts
@@ -1,21 +1,0 @@
-import { navigationActions } from '../../actions'
-import { ThunkAction } from '../../store'
-
-export {
-  consumeAuthenticationRequest,
-  sendAuthenticationResponse,
-} from './authenticationRequest'
-export {
-  consumeCredentialOfferRequest,
-  consumeCredentialReceive,
-  validateSelectionAndSave,
-} from './credentialOffer'
-export {
-  consumeCredentialRequest,
-  sendCredentialResponse,
-} from './credentialRequest'
-export { consumePaymentRequest, sendPaymentResponse } from './paymentRequest'
-
-export const cancelSSO: ThunkAction = dispatch => {
-  return dispatch(navigationActions.navigatorResetHome())
-}
