+++ src/lib/interactionManager/interaction.ts
@@ -25,7 +25,7 @@ import { Linking } from '../../polyfills/reactNative'
 import { AppError, ErrorCode } from '../errors'
 import { Authentication } from 'jolocom-lib/js/interactionTokens/authentication'
 import { Identity } from 'jolocom-lib/js/identity/identity'
-import { generateIdentitySummary } from '../../actions/sso/utils'
+import { generateIdentitySummary } from '../../utils/generateIdentitySummary'
 
 /***
  * - initiated by InteractionManager when an interaction starts
