--- src/reducers/registration/index.ts
@@ -1,15 +1,0 @@
-import { combineReducers } from 'redux'
-import { loading } from './loading'
-
-export interface LoadingState {
-  readonly loadingMsg: string
-  readonly isRegistering: boolean
-}
-
-export interface RegistrationState {
-  readonly loading: LoadingState
-}
-
-export const registrationReducer = combineReducers({
-  loading,
-})
