--- src/reducers/registration/loading.ts
@@ -1,27 +1,0 @@
-import { AnyAction } from 'redux'
-import { LoadingState } from './index'
-
-const initialState: LoadingState = {
-  loadingMsg: 'Loading',
-  isRegistering: false,
-}
-
-export const loading = (
-  state = initialState,
-  action: AnyAction,
-): LoadingState => {
-  switch (action.type) {
-    case 'SET_IS_REGISTERING':
-      return {
-        ...state,
-        isRegistering: action.value,
-      }
-    case 'SET_LOADING_MSG':
-      return {
-        ...state,
-        loadingMsg: action.value,
-      }
-    default:
-      return state
-  }
-}
